backend/requirements.txt
flask==2.2.5
werkzeug==2.2.3
celery==5.3.1
redis==5.0.1
python-dotenv
Pillow
